implemented a form of focus centered around the display center.

+++ include/crusta/FrustumVisibility.h
@@ -15,7 +15,7 @@ class FrustumVisibility : public VisibilityEvaluator
 {
 public:
     /** the specification of the viewing parameters */
-    GLFrustum<float> frustum;
+    GLFrustum<double> frustum;
     
 //- inherited from VisibilityEvaluator
 public:
